pyproject.toml
[project]
name = "ai-agents-service"
version = "0.0.1"
readme = "README.md"

requires-python = ">=3.12, <3.13"

dependencies = [
    "asyncpg ~=0.29.0",
    "duckduckgo-search>=7.3.2",
    "faiss-cpu ~=1.10.0",
    "fastapi ~=0.115.5",
    "genson ~=1.3.0",
    "httpx ~=0.27.2",
    "langchain-community ~=0.3.17",
    "langchain-core ~=0.3.34",
    "langchain-experimental ~=0.3.4",
    "langchain-groq ~=0.2.1",
    "langchain-huggingface ~=0.1.2",
    "langchain-ollama ~=0.2.3",
    "langchain-openai ~=0.2.9",
    "langgraph ~=0.2.70",
    "langgraph-checkpoint-sqlite ~=2.0.1",
    "langgraph-cli[inmem]",
    "langsmith ~=0.3.8",
    "matplotlib ~=3.10.0",
    "numexpr ~=2.10.1",
    "pandas ~=2.2.3",
    "psycopg[binary]",
    "pydantic ~=2.10.6",
    "pydantic-settings ~=2.7.1",
    "python-dotenv ~=1.0.1",
    "seaborn ~=0.13.2",
    "setuptools ~=75.8.0",
    "sqlalchemy",
    "streamlit ~=1.42.0",
    "tiktoken >=0.8.0",
    "typing-extensions ~=4.12.2",
    "unstructured[pdf,md,docx,pptx,xlsx,csv]",
    "uvicorn ~=0.34.0",
    "weaviate-client ~=4.10.4",
]

[dependency-groups]
client = [
    "httpx ~=0.27.2",
    "pydantic ~=2.10.6",
    "python-dotenv ~=1.0.1",
    "streamlit ~=1.42.0",
    "typing-extensions ~=4.12.2",
]

[tool.uv]
dev-dependencies = [
    "ruff",
]
